requirements.txt
# mongo
pymongo

# darksky
python-forecastio

# geo
geopy